Ad & Category Object + controller added
