+++ backend/http/category.http
@@ -1,5 +1,3 @@
-GET http://localhost:3000/Category
-###
 GET http://localhost:3000/Category?name=
 ###
 POST http://localhost:3000/Category
