--- backend/http/hello_world.http
@@ -1,1 +1,0 @@
-GET http://localhost:3000/
